Move CHPL populator code to a package so it can be used elsewhere in the project
